# docs/api/API 設計書（OpenAPI 3.0.3）.yaml
openapi: 3.0.3
info:
  title: 駐車場予約システム API
  version: 0.1.0
  description: >
    MVP の HTTP API 契約。バックエンド処理一覧の A 群（App Service ハンドラ）に対応。
    時刻はすべて UTC（RFC 3339, 例 2026-06-18T09:00:00Z）。認証は Bearer JWT。
    金額は最小通貨単位の数値（JPY は円）。
    エラーは Error スキーマで返し、retryable で再試行可否を、code で詳細理由を表す。
servers:
  - url: https://{host}/api
    description: App Service
    variables:
      host:
        default: example.azurewebsites.net
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: 会員登録
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: 登録成功（トークン発行）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '409':
          description: メールアドレス重複
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [auth]
      summary: ログイン
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 認証成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: アカウント無効（退会済み等。Error.code=account_disabled）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: レート制限・アカウントロック（Error.code=account_locked）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: トークン再発行
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: 再発行成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/logout:
    post:
      tags: [auth]
      summary: ログアウト（リフレッシュトークンを失効）
      description: Bearer のアクセストークンで認証し、ボディの refresh_token を失効する（MVP は access denylist 無し）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204': { description: 成功 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: 入力検証エラー
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            機械可読の理由コード。例：physical_occupancy / device_unhealthy /
            conflict_overlap / conflict_buffer / timeout / validation_error /
            unauthorized / email_conflict / invalid_credentials / account_locked /
            account_disabled / invalid_token / token_reused / internal_error
          example: physical_occupancy
        message: { type: string, example: 区画が使用中です }
        retryable:
          type: boolean
          description: 再試行で解消し得るか（504=true、409 物理占有/競合=false など）
        retry_after:
          type: integer
          nullable: true
          description: 再試行までの推奨待機秒数（任意）
    RegisterRequest:
      type: object
      required: [email, password, terms_version]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        name: { type: string, nullable: true }
        terms_version: { type: string, example: "2026-01" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, description: アクセストークンの有効秒数, example: 900 }
...